interface1/interface_core/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "interface-core"
version = "0.1.0"
description = "Autonomous middleware glue daemon — proactive discovery, library consolidation, and dynamic bridging between disparate software nodes."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Interface Core" }]
keywords = ["middleware", "bridge", "autonomous", "glue", "integration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: System :: Networking",
]
dependencies = [
    "pydantic>=1.10,<2.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
full = [
    "asyncio-mqtt>=0.16.0",
    "aiokafka>=0.10.0",
    "pyzmq>=25.0",
    "confluent-kafka>=2.3.0",
    "python-consul2>=0.1.5",
    "sentence-transformers>=2.7.0",
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp>=1.24.0",
    "prometheus_client>=0.20.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
interface-core = "interface_core.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["interface_core*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
